spike/series0ne/erp-oauth Final implementation changes for OAuth and TIX integration As there are a lot of changes in this commit, please ask me if you are unsure of anything.

+++ tradeix-concord-shared-client/src/main/kotlin/com/tradeix/concord/shared/client/components/TIXConfiguration.kt
@@ -13,5 +13,5 @@ class TIXConfiguration(
         @Value("\${config.tix.webapi.url}") val webApiUrl: String,
         @Value("\${config.tix.webapi.composer_program_id}") val composerProgramId: String,
         @Value("\${config.tix.webapi.on_behalf_of}") val onBehalfOf: String,
-        @Value("\${config.tix.vault_observer_timeout}") val vaultObserverTimeout: Int
+        @Value("\${config.tix.vault_observer_timeout}") val vaultObserverTimeout: Long
 )
